feat: Unify element editing operations
