+++ src/renderer/src/i18n/en.ts
@@ -390,6 +390,9 @@ Business presentations, product introductions, project summaries — formal but 
     exitAdjust: 'Exit adjustment',
     adjustLayout: 'Adjust position/size',
     saveAdjustments: 'Save adjustments',
+    deleteElement: 'Delete element',
+    deleteElementConfirm: 'Are you sure you want to delete this element? This cannot be undone.',
+    deleteElementFailed: 'Failed to delete element',
     exitInspect: 'Exit inspect',
     inspectElement: 'Inspect element',
     aiMode: 'AI',
