+++ src/renderer/src/i18n/zh.ts
@@ -374,6 +374,9 @@ export const zh = {
     exitAdjust: '退出调整',
     adjustLayout: '调整位置/大小',
     saveAdjustments: '保存调整',
+    deleteElement: '删除元素',
+    deleteElementConfirm: '确定删除该元素吗？删除后将无法恢复。',
+    deleteElementFailed: '删除元素失败',
     exitInspect: '退出检选',
     inspectElement: '检选元素',
     aiMode: 'AI 模式',
